feat: Add promotion email delivery functionality
The code changes introduce a new function `deliver_promotion` in the `Pento.Accounts.UserNotifier` module to send a promotion email to a user. This function is called in the `Pento.Promo.send_promo` function in the `Pento.Promo` module.

Based on the recent user commits and repository commits, the suggested commit message is "feat: Add promotion email delivery functionality".