feat: Save rating in rating form live view
This commit adds the `save_rating` function to the `PentoWeb.RatingLive.Form` module. The function is called when the "save" event is triggered and it saves the rating using the `Survey.create_rating` function. If the rating is successfully created, it updates the product's ratings and sends a `:created_rating` message. If there is an error with the changeset, it assigns the changeset to the socket.

This code change enhances the functionality of the rating form live view and improves the user experience by allowing users to save their ratings.